Added stackoverflow_urls and unittests
stackoverflow_urls is a _util function to check whether various SO urls are
in a string. It uses regex to search through the string and returns a dict with
4 key/value pairs corresponding to:

- Question urls
- Answer urls
- Comment urls
- User urls (i.e. profile page).

It returns these urls inside lists inside a dict. I kept them as lists even though
that means there could be multiple urls for the same question. These can always
be tidied up further on in the analysis.

I've written some (what I think :P) extensive unittests for the function.